Updated the OLED SSD1306 Demos and Drivers with Basic Functionality to work with latest nF
Updated the code to make minimal changes to make it work with OLED. Tested with 128x64 and 128x32 and it works fine.

Updated Font logic and created basic Font logic and drawing on screen.

There is more work to do, but this is initial update to make the code work, and see what direction will need to be taken for the other controllers/devices if they will be supported.