Improved sensor read stability
We now sleep the expected amount of time it takes to receive a
response as indicated by the documentation of the SPS30. Also
removed an unused MemoryStream. Simplified reading by no longer
trying to read as many frames as we can, but instead we clear the
bytes in the serial buffer before we begin.