Fix ReadSingle*() in Release build (#439)
In release builds, and also only when the result of the
ReadSingle*() was not assigned to a float (but anything else, like
double), the assigned value would be its underlying uint value.

A forced assignment in the ReadSingle*() functions fixes this
problem for release builds. The test has been updated to capture
this edge case.