Fix NullReferenceException bug in AtChannel.ReaderLoopAsync
Right after the lock in AtChannel.ReaderLoopAsync it is checked if
_currentCommand != null. After that _currentCommand is used to do other
checks.
Sometimes these last checks give a NullReferenceException because
another thread sets _currentCommand to null in the 'finally' of
AtChannel.SendFullCommand. This sometimes happens after an hour or so..

To fix this we can put a lock around the code in the 'finally' of
AtChannel.SendFullCommand so the checking and using and the setting to
null are mutually exclusive and the NullReferenceException doesn't
happen.

Signed-off-by: Rik Kreeftenberg <rik.kreeftenberg@spinsoft.nl>

#1579