Fix hanging bug in Sim7672HttpClient.SendInternal
When the response is processed at the end of Sim7672HttpClient.SendInternal the idea was to limit the time to 2 minutes.

Because of a logical bug this time is unlimited so when something unexpected happens the system continues to wait indefinitely.

The while loop should read:
While there is something left to read AND Cancelation is NOT requested
DO

Also, when the line simply reads "ERROR" then you can stop reading and break the While loop. Otherwise it will wait for 2 minutes. (Or indefinitely if the previous bug is not fixed)

Signed-off-by: Rik Kreeftenberg <rik.kreeftenberg@spinsoft.nl>

#1579