Bugfix HTTPINIT Retry mechanism Sim7672HttpClient.SendInternal
The first thing that is necessary to send a HTTP message is to send the "AT+HTTPINIT" command.
In the current situation when this fails, it will try it one more time and when that fails it just continues as if nothing is wrong..

The unused "Retry:" and the "int retries = 5" variable suggest that there once was a retry mechanism and in the Sim800HttpClient there is.

So I copied this retry mechanism to the Sim7672HttpClient.SendInternal, including the return of "ServiceUnavailable" when it fails after 5 retries.

I haven't noticed this gone wrong but the original implementation suggests it might. And after this fix it is handled correctly.

Signed-off-by: Rik Kreeftenberg <rik.kreeftenberg@spinsoft.nl>

#1579