[AtModem]Fix ManualResetEvent bug in AtChannel.SendFullCommand
The _waitingForCommandResponse ManualResetEvent is created when the AtChannel is created. It is used to halt the return of ManualResetEvent until the _currentResponse is complete.

It's not necessary to create ManualResetEvent every time AtChannel.SendFullCommand is called.

To fix this you should not create a new ManualresetEvent but simply call Reset on it. Just like it's done in Sim7672HttpClient.

#1579