[AtModem] Fix NullReferenceException in AtChannel.
The _currentCommand and the _currentResponse are used troughout the AtChannel class. They are set to null in the 'finally' of the AtChannel.SendFullCommand.
But in very few occasions (once every hour or so) the _currentCommand is used by another thread after it has been set to null, resulting in a NullReferenceException. For example in HandleFinalResponse.
This is the same multi-threading problem like the one fixed in PR #1216.

We can fix this by placing a lock around the call to AtChannel.ProcessMessage so access to _currentCommand and _currentresponse is mutually exclusive to setting them to null. So NullReference cannot occur.

Signed-off-by: Rik Kreeftenberg <rik.kreeftenberg@spinsoft.nl>

#1579